+++ test/fixtures/card-with-flavor-name-on-multiple-faces.ts
@@ -1,0 +1,159 @@
+module.exports = {
+  object: "card",
+  id: "83a86763-c3fe-4cff-8b65-22cb329d112c",
+  oracle_id: "e71ac446-02a4-4468-8d29-f28b21617665",
+  multiverse_ids: [],
+  tcgplayer_id: 252008,
+  cardmarket_id: 611222,
+  name: "Havengul Laboratory // Havengul Mystery",
+  lang: "en",
+  released_at: "2021-10-18",
+  uri: "https://api.scryfall.com/cards/83a86763-c3fe-4cff-8b65-22cb329d112c",
+  scryfall_uri:
+    "https://scryfall.com/card/sld/609/havengul-laboratory-havengul-mystery?utm_source=api",
+  layout: "transform",
+  highres_image: true,
+  image_status: "highres_scan",
+  cmc: 0.0,
+  type_line: "Legendary Land // Legendary Land",
+  color_identity: ["B"],
+  keywords: ["Investigate", "Transform"],
+  produced_mana: ["B", "C"],
+  card_faces: [
+    {
+      object: "card_face",
+      name: "Havengul Laboratory",
+      flavor_name: "Hawkins National Laboratory",
+      printed_name: "Hawkins National Laboratory",
+      mana_cost: "",
+      type_line: "Legendary Land",
+      oracle_text:
+        "{T}: Add {C}.\n{4}, {T}: Investigate.\nAt the beginning of your end step, if you sacrificed three or more Clues this turn, transform Havengul Laboratory.",
+      colors: [],
+      flavor_text:
+        '"Today, we make history. Today, we make contact."\n—Dr. Martin Brenner',
+      artist: "Ravenna Tran",
+      artist_id: "e24bc1d0-446b-45e0-b215-89f581837aa4",
+      illustration_id: "57eb443c-e039-4988-9f3c-e3f311667d12",
+      image_uris: {
+        small:
+          "https://c1.scryfall.com/file/scryfall-cards/small/front/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+        normal:
+          "https://c1.scryfall.com/file/scryfall-cards/normal/front/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+        large:
+          "https://c1.scryfall.com/file/scryfall-cards/large/front/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+        png: "https://c1.scryfall.com/file/scryfall-cards/png/front/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.png?1646349812",
+        art_crop:
+          "https://c1.scryfall.com/file/scryfall-cards/art_crop/front/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+        border_crop:
+          "https://c1.scryfall.com/file/scryfall-cards/border_crop/front/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+      },
+    },
+    {
+      object: "card_face",
+      name: "Havengul Mystery",
+      flavor_name: "The Upside Down",
+      printed_name: "The Upside Down",
+      mana_cost: "",
+      type_line: "Legendary Land",
+      oracle_text:
+        "When this land transforms into Havengul Mystery, return target creature card from your graveyard to the battlefield.\nWhen the creature put onto the battlefield with Havengul Mystery leaves the battlefield, transform Havengul Mystery.\n{T}, Pay 1 life: Add {B}.",
+      colors: [],
+      artist: "Ravenna Tran",
+      artist_id: "e24bc1d0-446b-45e0-b215-89f581837aa4",
+      illustration_id: "9fa64243-7bee-4238-9739-fd7b0bf73c72",
+      image_uris: {
+        small:
+          "https://c1.scryfall.com/file/scryfall-cards/small/back/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+        normal:
+          "https://c1.scryfall.com/file/scryfall-cards/normal/back/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+        large:
+          "https://c1.scryfall.com/file/scryfall-cards/large/back/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+        png: "https://c1.scryfall.com/file/scryfall-cards/png/back/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.png?1646349812",
+        art_crop:
+          "https://c1.scryfall.com/file/scryfall-cards/art_crop/back/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+        border_crop:
+          "https://c1.scryfall.com/file/scryfall-cards/border_crop/back/8/3/83a86763-c3fe-4cff-8b65-22cb329d112c.jpg?1646349812",
+      },
+    },
+  ],
+  legalities: {
+    standard: "not_legal",
+    future: "not_legal",
+    historic: "not_legal",
+    gladiator: "not_legal",
+    pioneer: "not_legal",
+    modern: "not_legal",
+    legacy: "legal",
+    pauper: "not_legal",
+    vintage: "legal",
+    penny: "not_legal",
+    commander: "legal",
+    brawl: "not_legal",
+    historicbrawl: "not_legal",
+    alchemy: "not_legal",
+    paupercommander: "not_legal",
+    duel: "legal",
+    oldschool: "not_legal",
+    premodern: "not_legal",
+  },
+  games: ["paper"],
+  reserved: false,
+  foil: true,
+  nonfoil: true,
+  finishes: ["nonfoil", "foil"],
+  oversized: false,
+  promo: false,
+  reprint: false,
+  variation: false,
+  set_id: "4d92a8a7-ccb0-437d-abdc-9d70fc5ed672",
+  set: "sld",
+  set_name: "Secret Lair Drop",
+  set_type: "box",
+  set_uri: "https://api.scryfall.com/sets/4d92a8a7-ccb0-437d-abdc-9d70fc5ed672",
+  set_search_uri:
+    "https://api.scryfall.com/cards/search?order=set&q=e%3Asld&unique=prints",
+  scryfall_set_uri: "https://scryfall.com/sets/sld?utm_source=api",
+  rulings_uri:
+    "https://api.scryfall.com/cards/83a86763-c3fe-4cff-8b65-22cb329d112c/rulings",
+  prints_search_uri:
+    "https://api.scryfall.com/cards/search?order=released&q=oracleid%3Ae71ac446-02a4-4468-8d29-f28b21617665&unique=prints",
+  collector_number: "609",
+  digital: false,
+  rarity: "rare",
+  artist: "Ravenna Tran",
+  artist_ids: ["e24bc1d0-446b-45e0-b215-89f581837aa4"],
+  border_color: "black",
+  frame: "2015",
+  frame_effects: ["legendary"],
+  security_stamp: "triangle",
+  full_art: false,
+  textless: false,
+  booster: false,
+  story_spotlight: false,
+  prices: {
+    usd: "14.65",
+    usd_foil: "14.72",
+    usd_etched: null,
+    eur: "13.00",
+    eur_foil: "12.70",
+    tix: null,
+  },
+  related_uris: {
+    tcgplayer_infinite_articles:
+      "https://infinite.tcgplayer.com/search?contentMode=article&game=magic&partner=scryfall&q=Havengul+Laboratory+%2F%2F+Havengul+Mystery&utm_campaign=affiliate&utm_medium=api&utm_source=scryfall",
+    tcgplayer_infinite_decks:
+      "https://infinite.tcgplayer.com/search?contentMode=deck&game=magic&partner=scryfall&q=Havengul+Laboratory+%2F%2F+Havengul+Mystery&utm_campaign=affiliate&utm_medium=api&utm_source=scryfall",
+    edhrec: "https://edhrec.com/route/?cc=Havengul+Laboratory",
+    mtgtop8:
+      "https://mtgtop8.com/search?MD_check=1&SB_check=1&cards=Havengul+Laboratory",
+  },
+  purchase_uris: {
+    tcgplayer:
+      "https://shop.tcgplayer.com/product/productsearch?id=252008&utm_campaign=affiliate&utm_medium=api&utm_source=scryfall",
+    cardmarket:
+      "https://www.cardmarket.com/en/Magic/Products/Search?referrer=scryfall&searchString=Havengul+Laboratory&utm_campaign=card_prices&utm_medium=text&utm_source=scryfall",
+    cardhoarder:
+      "https://www.cardhoarder.com/cards?affiliate_id=scryfall&data%5Bsearch%5D=Havengul+Laboratory&ref=card-profile&utm_campaign=affiliate&utm_medium=card&utm_source=scryfall",
+  },
+};